update MarketingCheckbox to details summary pattern
